check if requested userid is exact from token

+++ userAPI/persuasiveAPI/Controllers/WishlistController.cs
@@ -10,15 +10,22 @@ namespace persuasiveAPI.Controllers;
 public class WishlistController: ControllerBase
 {
     private readonly IWishlistRepository _wishlistRepository;
-    public WishlistController(IWishlistRepository wishlistRepository)
+    private readonly IJwtTokenService _jwtTokenService;
+    public WishlistController(IWishlistRepository wishlistRepository, IJwtTokenService jwtTokenService)
     {
         _wishlistRepository = wishlistRepository;
+        _jwtTokenService = jwtTokenService;
     }
     [HttpGet]
     [Route("{userId}")]
     [Authorize(Roles = UserRoles.User)]
     public async Task<IActionResult> GetWishlist(string userId)
     {
+        var token = Request.Headers["Authorization"].ToString().Substring(7);
+        var isValid = _jwtTokenService.CheckAccessTokenWithUserId(token, userId);
+        if(!isValid)
+            return Unauthorized("Requested user id does not match token");
+
         var wishlist = await _wishlistRepository.GetWishlistByUserId(userId);
         return Ok(wishlist);
     }
@@ -27,6 +34,11 @@ public class WishlistController: ControllerBase
     [Authorize(Roles = UserRoles.User)]
     public async Task<IActionResult> AddProduct(string userId, int productId)
     {
+        var token = Request.Headers["Authorization"].ToString().Substring(7);
+        var isValid = _jwtTokenService.CheckAccessTokenWithUserId(token, userId);
+        if(!isValid)
+            return Unauthorized("Requested user id does not match token");
+
         await _wishlistRepository.AddProductToWishlist(userId, productId);
         return Ok();
     }
@@ -35,6 +47,11 @@ public class WishlistController: ControllerBase
     [Authorize(Roles = UserRoles.User)]
     public async Task<IActionResult> RemoveProduct(string userId, int productId)
     {
+        var token = Request.Headers["Authorization"].ToString().Substring(7);
+        var isValid = _jwtTokenService.CheckAccessTokenWithUserId(token, userId);
+        if(!isValid)
+            return Unauthorized("Requested user id does not match token");
+
         await _wishlistRepository.RemoveProductFromWishlist(userId, productId);
         return Ok();
     }
